refactor: LenmaTemplateManager
