+++ loglead/parsers/__init__.py
@@ -3,5 +3,6 @@ from .bert.bertembedding import BertEmbeddings
 from .Brain.Brain import BrainParser
 # drain3
 from .iplom.IPLoM import IPLoMParser
+from .lenma.lenma import LenmaTemplateManager
 
-__all__ = ['AELParser', 'BertEmbeddings', 'BrainParser', 'IPLoMParser']
+__all__ = ['AELParser', 'BertEmbeddings', 'BrainParser', 'IPLoMParser', 'LenmaTemplateManager']
